Color management: Tweak compatible look check
Change compatible look check so that if the view contains
explicitly configured looks the non-explicit looks are not
added to the list.

In practice this means that if there are looks "Low Contrast"
and "AgX - Low Contrast" only the latter one is considered to
be compatible with the "AgX" view.

Ref #110685

Pull Request: https://projects.blender.org/blender/blender/pulls/111229