Fix #111504: Regression: Mesh draw corruption in sculpt mode
The commit f10965dcb8 introduced a regression in the way how the VBOs
are filled by assuming the requested VBO attribute type has the same
alignment as the CPU and is the same on different platforms.

Unfortunately, this turned out to not be the case.

Switch the mask attribute to be float on the GPU, which has a downside
of increased bandwidth to be transferred, but a benefit of less compute
power needed to update the VBO.

The fix is suggested by Clement.

Pull Request: https://projects.blender.org/blender/blender/pulls/111521